Moved constants related utility functions into core.constants
