+++ virtual_rainforest/models/abiotic_simple/abiotic_simple_model.py
@@ -24,9 +24,10 @@ from pint import Quantity
 from xarray import DataArray
 
 from virtual_rainforest.core.base_model import BaseModel
+from virtual_rainforest.core.constants import load_constants
 from virtual_rainforest.core.data import Data
 from virtual_rainforest.core.logger import LOGGER
-from virtual_rainforest.core.utils import load_constants, set_layer_roles
+from virtual_rainforest.core.utils import set_layer_roles
 from virtual_rainforest.models.abiotic_simple import microclimate
 from virtual_rainforest.models.abiotic_simple.constants import AbioticSimpleConsts
