+++ virtual_rainforest/models/soil/soil_model.py
@@ -27,10 +27,11 @@ from scipy.integrate import solve_ivp
 from xarray import DataArray
 
 from virtual_rainforest.core.base_model import BaseModel
+from virtual_rainforest.core.constants import load_constants
 from virtual_rainforest.core.data import Data
 from virtual_rainforest.core.exceptions import InitialisationError
 from virtual_rainforest.core.logger import LOGGER
-from virtual_rainforest.core.utils import load_constants, set_layer_roles
+from virtual_rainforest.core.utils import set_layer_roles
 from virtual_rainforest.models.soil.carbon import calculate_soil_carbon_updates
 from virtual_rainforest.models.soil.constants import SoilConsts
